scripts/upwork_scraper_c/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-dotenv==1.0.0
fake-useragent==1.4.0
webdriver-manager==4.0.1 